backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.8.0
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.141.0
google-cloud-firestore==2.20.0
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
python-multipart==0.0.6
cors==1.0.1
python-jose[cryptography]==3.3.0
